Move Admin's settings to sub-namespace
