+++ src/Core/class-Init.php
@@ -101,7 +101,7 @@ if ( ! class_exists( Init::class ) ) {
 			$this->loader->add_action( 'admin_enqueue_scripts', $assets, 'enqueue_styles' );
 			$this->loader->add_action( 'admin_enqueue_scripts', $assets, 'enqueue_scripts' );
 
-			$settings = new Admin\Settings();
+			$settings = new Admin\Settings\Main();
 
 			// Plugin action links
 			$this->loader->add_filter( 'plugin_action_links_' . Plugin_Data::plugin_basename(), $settings, 'add_action_links' );
